+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Is64Bit Utility property (credit @rasta-mouse)
 - Added Is64BitProcess Host function (credit @TheWover)
 - Added GetProcessorArchitecture, GetParentProcess, GetProcessOwner, IsWow64, and supporting P/Invoke signatures (credit @rasta-mouse)
+- Added Keylogger class (credit @checkymander)
 
 ### Changed
 - Improved DynamicInvoke library (credit @TheWover)